requirements.txt
Django==1.9.3
Twisted==15.5.0
argparse==1.2.1
asgi-redis==0.8.3
asgiref==0.9
autobahn==0.12.1
channels==0.9.3
daphne==0.9.1
dj-database-url==0.4.0
django-allauth==0.28.0
django-crispy-forms==1.6.1
haikunator==1.0.1
msgpack-python==0.4.7
oauthlib==2.0.0
psycopg2==2.6.1
python-openid==2.2.5
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
txaio==2.2.1
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.1.3
